[refactor] params.ts 추가
