+++ src/types/params.ts
@@ -1,0 +1,3 @@
+export interface MatchParams {
+  roomPK: string
+};
